+++ ci/cpu/cugraph/build_cugraph.sh
@@ -1,18 +1,9 @@
 #!/usr/bin/env bash
 set -e
 
-# If nightly build, append current YYMMDD to version
-if [[ "$BUILD_MODE" = "branch" && "$SOURCE_BRANCH" = branch-* ]] ; then
-  export VERSION_SUFFIX=`date +%y%m%d`
-fi
-
 if [ "$BUILD_CUGRAPH" == "1" ]; then
   echo "Building cugraph"
   CUDA_REL=${CUDA_VERSION%.*}
 
-  if [ "$BUILD_ABI" == "1" ]; then
-    conda build conda/recipes/cugraph -c nvidia/label/cuda${CUDA_REL} -c rapidsai/label/cuda${CUDA_REL} -c rapidsai-nightly/label/cuda${CUDA_REL} -c numba -c conda-forge -c defaults --python=$PYTHON
-  else
-    conda build conda/recipes/cugraph -c rapidsai/label/cf201901 -c nvidia/label/cf201901 -c numba -c conda-forge/label/cf201901 -c defaults --python=$PYTHON
-  fi
+  conda build conda/recipes/cugraph --python=$PYTHON
 fi